Two attempts to optimize locators.
Step one is to allow a locator to give back more precise data than
requested.  This allows to compute with larger numbers better, as it
avoids "order of magnitude" search.

Step two is an attempt to memoize outputs of locators.  This did not
seem to change performance much, or if anything, make it
worse.  Further experiments and smart designs are necessary.

Note: these are very crude developments that would need to be
rewritten in a next iteration.